Refactor code , use prosemirror to add classes to deletion decoration.
